refactor: Rename `lib/serializeError.js`
