+++ lib/error-handler.js
@@ -5,7 +5,7 @@
  */
 
 const log = require('./log');
-const serializeError = require('./serializeError');
+const serializeError = require('./serialize-error');
 const { getDashboardUrl } = require('./dashboard');
 const { parseDeploymentData } = require('./deployment');
