+++ lib/serialize-error.js
